Master update : - BMS Slave hardware v2 - BMS Slave software v2 - BMS Master hardware v2 (was already on the SVN however) - BMS Master software v2
This is a major update pushing the last version of hardware and software of the BMS BMSafe project. There have been a lots of improvements in the software, re-organization of functions, clarification of the code, usage of a library for CAN communication, new functionalities, etc. The BMS circuits have been tested and most of their functionalities work properly. There are still certain known bugs to be corrected. We will try to make more updates as we correct these.

The subsequent updates will mostly be for documentation. There is a big lack in documentation for this project, so we are planning to add some. We already started documenting part of the code using Doxygen compatible syntax. We will probably publish the auto-generated documentation in the following updates.